동영상 API
1. 동영상 업로드
2. 동영상 조회
3. 동영상 북마크
4. 동영상 스트리밍

+++ src/main/java/com/kong/backend/service/AlertService.java
@@ -196,6 +196,6 @@ public class AlertService {
         private Double layRate;            // nullable
         private Double prob;               // nullable
         private Double ts;                 // nullable (epoch seconds, float)
-        private Long videoId;              // nullable (VideoEntity FK)
+        private Integer videoId;              // nullable (VideoEntity FK)
     }
 }
